refactor: move Walker's algorithm to sub dir
